Fix GameBoard generation GameBoard generation tests are now passing papa bless

+++ sneik/sneikTest/ClassesTest/GameBoardTest.cs
@@ -15,35 +15,35 @@ namespace sneikTest.ClassesTest
         {
             Difficulty difficulty = Difficulty.EASY;
             GameBoard gameBoard = new GameBoard(difficulty);
-            Assert.AreEqual(ObstacleCoef.EASY,gameBoard.Obstacles.Count);
+            Assert.AreEqual((int)ObstacleCoef.EASY,gameBoard.Obstacles.Count);
         }
         [TestMethod]
         public void CheckGameboardMedium()
         {
             Difficulty difficulty = Difficulty.MEDIUM;
             GameBoard gameBoard = new GameBoard(difficulty);
-            Assert.AreEqual(ObstacleCoef.MEDIUM, gameBoard.Obstacles.Count);
+            Assert.AreEqual((int)ObstacleCoef.MEDIUM, gameBoard.Obstacles.Count);
         }
         [TestMethod]
         public void CheckGameboardHard()
         {
             Difficulty difficulty = Difficulty.HARD;
             GameBoard gameBoard = new GameBoard(difficulty);
-            Assert.AreEqual(ObstacleCoef.HARD, gameBoard.Obstacles.Count);
+            Assert.AreEqual((int)ObstacleCoef.HARD, gameBoard.Obstacles.Count);
         }
         [TestMethod]
         public void CheckGameboardVeryHard()
         {
             Difficulty difficulty = Difficulty.VERY_HARD;
             GameBoard gameBoard = new GameBoard(difficulty);
-            Assert.AreEqual(ObstacleCoef.VERY_HARD, gameBoard.Obstacles.Count);
+            Assert.AreEqual((int)ObstacleCoef.VERY_HARD, gameBoard.Obstacles.Count);
         }
         [TestMethod]
         public void CheckGameboardNightmare()
         {
             Difficulty difficulty = Difficulty.NIGHTMARE;
             GameBoard gameBoard = new GameBoard(difficulty);
-            Assert.AreEqual(ObstacleCoef.NIGHTMARE, gameBoard.Obstacles.Count);
+            Assert.AreEqual((int)ObstacleCoef.NIGHTMARE, gameBoard.Obstacles.Count);
         }
     }
 }
